--- controller/categoryController.js
@@ -1,12 +1,0 @@
-const categoryService = require('../service/categoryService');
-
-function getCategories(req, res, next) {
-    return categoryService.getCategories().then((val) => {
-        res.send(val);
-    });
-}
-
-
-module.exports = {
-    getCategories: getCategories
-};
