+++ controller/userController.js
@@ -1,0 +1,11 @@
+const getUsersService = require('../service/userService');
+
+function getUsers(req, res, next) {
+    return getUsersService.getUsers().then((val) => {
+        res.send(val);
+    });
+}
+
+module.exports = {
+    getUsers: getUsers
+};
