+++ models/user.js
@@ -1,0 +1,9 @@
+module.exports = (sequelize, DataTypes) => {
+    return sequelize.define('user', {
+        name: {
+            type: DataTypes.STRING(20),
+            allowNull: false,
+            unique: false
+        }
+    });
+};
