+++ routes/userRouter.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+const userController = require('../controller/userController');
+
+router.get('/', userController.getUsers);
+
+
+module.exports = router;
