--- service/categoryService.js
@@ -1,9 +1,0 @@
-const categoryRepository = require('../persistence/categoryRepository');
-
-function getCategories() {
-    return categoryRepository.getCategories();
-}
-
-module.exports = {
-    getCategories: getCategories
-};
