+++ service/userService.js
@@ -1,0 +1,9 @@
+const userRepository = require('../persistence/userRepository');
+
+function getUsers() {
+    return userRepository.getUsers();
+}
+
+module.exports = {
+    getUsers: getUsers
+};
